requirements.txt
yt-dlp
openai
anthropic
python-dotenv
